# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "version": "1.0",
        "contact": {}
    },
    "host": "texasholdem.com",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/token/refresh": {
            "post": {
                "tags": [
                    "/auth - 授权相关 "
                ],
                "summary": "刷新token",
                "description": "刷新token",
                "operationId": ".刷新token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "post 数据",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/c_auth.TokenRefreshParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/c_auth.TokenRefreshResult"
                        }
                    }
                }
            }
        },
...
        "/user/{user_id}/bag/use": {
            "get": {
                "tags": [
                    "/user/:user_id - 用户相关 "
                ],
                "summary": "用户背包物品使用",
                "description": "用户背包物品使用",
                "operationId": ".用户背包物品使用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "用户id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "query",
                        "name": "token",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "goods_id",
                        "description": "goods_id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/define.BaseResult"
                        }
                    }
                }
            }
        },
...
        "/user/{user_id}/detail": {
            "get": {
                "tags": [
                    "/user/:user_id - 用户相关 "
                ],
                "summary": "获取用户详细信息",
                "description": "获取用户详细信息",
                "operationId": ".获取用户详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "token",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "user_id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/c_user.GetUserInfoDetailResult"
                        }
                    }
                }
            }
        },
...
        "/user/{user_id}/info": {
            "get": {
                "tags": [
                    "/user/:user_id - 用户相关 "
                ],
                "summary": "获取用户基本信息",
                "description": "获取用户基本信息",
                "operationId": ".获取用户基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "token",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "path",
                        "name": "user_id",
                        "description": "user_id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/c_user.UserInfoResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "c_auth.TokenRefreshParams": {
            "title": "TokenRefreshParams",
            "type": "object",
            "properties": {
                "from": {
                    "description": "请求来源",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "旧的刷新码",
                    "type": "string"
                },
                "token": {
                    "description": "旧的token",
                    "type": "string"
                },
                "unique_id": {
                    "description": "唯一设备",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "c_auth.TokenRefreshResult": {
            "title": "LoginResult",
            "type": "object",
            "properties": {
                "expire": {
                    "description": "token过期时间",
                    "type": "integer",
                    "format": "int64"
                },
                "fresh": {
                    "description": "是否是新用户",
                    "type": "integer",
                    "format": "int64"
                },
                "msg": {
                    "type": "string"
                },
                "need_captcha": {
                    "description": "是否需要验证码",
                    "type": "boolean"
                },
                "refresh_token": {
                    "description": "当 Token 过期时，使用该 Token 获取一个新的",
                    "type": "string"
                },
                "ret": {
                    "description": "1成功 0失败",
                    "type": "integer",
                    "format": "int64"
                },
                "token": {
                    "description": "请求授权的token",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户Id",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
        "c_user.GetUserInfoDetailResult": {
            "title": "GetUserInfoDetailResult",
            "type": "object",
            "properties": {
                "best_winner": {
                    "description": "最大赢取筹码",
                    "type": "integer",
                    "format": "int64"
                },
                "cards": {
                    "description": "最大牌",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Card"
                    }
                },
                "inbound_rate": {
                    "description": "入局率",
                    "type": "number",
                    "format": "double"
                },
                "msg": {
                    "type": "string"
                },
                "ret": {
                    "description": "1成功 0失败",
                    "type": "integer",
                    "format": "int64"
                },
                "showdown_rate": {
                    "description": "摊牌率",
                    "type": "number",
                    "format": "double"
                },
                "total_game": {
                    "description": "总局数",
                    "type": "integer",
                    "format": "int64"
                },
                "win_rate": {
                    "description": "胜率",
                    "type": "number",
                    "format": "double"
                }
            }
        },
...
        "c_user.UserInfo": {
            "title": "UserInfo",
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "balance": {
                    "description": "金币余额",
                    "type": "integer",
                    "format": "int64"
                },
                "checkin_days": {
                    "description": "签到天数",
                    "type": "integer",
                    "format": "int64"
                },
                "diamond_balance": {
                    "description": "钻石余额",
                    "type": "integer",
                    "format": "int64"
                },
                "gender": {
                    "description": "性♂别 // 0 未知  1男2女 ",
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "is_chekin": {
                    "description": "今天是否签到,,1已签0未签",
                    "type": "integer",
                    "format": "int64"
                },
                "is_fresh": {
                    "description": "是否新用户",
                    "type": "integer",
                    "format": "int64"
                },
                "last_checkin_time": {
                    "description": "上次签到时间",
                    "type": "string"
                },
                "mobile_number": {
                    "description": "手机号",
                    "type": "string"
                },
                "nick_name": {
                    "description": "昵称",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "integer",
                    "format": "int64"
                },
                "type": {
                    "description": "用户类型",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "c_user.UserInfoResult": {
            "title": "UserInfoResult",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/c_user.UserInfo"
                },
                "msg": {
                    "type": "string"
                },
                "ret": {
                    "description": "1成功 0失败",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "define.BaseResult": {
            "title": "BaseResult",
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "ret": {
                    "description": "1成功 0失败",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "models.Card": {
            "title": "Card",
            "type": "object",
            "properties": {
                "suit": {
                    "description": "花色",
                    "type": "integer",
                    "format": "int64"
                },
                "value": {
                    "description": "牌值",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    }
}